port mimic3 provider to new synth api

+++ providers/mimic3/speech_provider_mimic3/main.py
@@ -30,38 +30,42 @@ class Mimic3SynthWorker(GObject.Object):
         # create the elements
         self.source = None
         self.parse = Gst.ElementFactory.make("rawaudioparse", "parse")
-        self.pitch = Gst.ElementFactory.make("pitch", "pitch")
         self.convert = Gst.ElementFactory.make("audioconvert", "convert")
-        self.sink = Gst.ElementFactory.make("autoaudiosink", "sink")
+        self.pitch = Gst.ElementFactory.make("pitch", "pitch")
+        self.convert2 = Gst.ElementFactory.make("audioconvert", "convert2")
+        self.caps_filter = Gst.ElementFactory.make("capsfilter", "audioconvert_filter")
+        self.sink = Gst.ElementFactory.make("fdsink", "sink")
         self.sink.set_property("sync", False)
 
-        # build the pipeline.
-        self.pipeline.add(self.sink)
-        self.pipeline.add(self.pitch)
-        self.pipeline.add(self.convert)
-        self.pipeline.add(self.parse)
+        elements = [
+            self.parse,
+            self.convert,
+            self.pitch,
+            self.convert2,
+            self.caps_filter,
+            self.sink,
+        ]
 
-        self.parse.link(self.convert)
-        self.convert.link(self.pitch)
-        self.pitch.link(self.sink)
+        for el in elements:
+            self.pipeline.add(el)
 
-        self.parse.set_property("sample-rate", SAMPLE_RATE)
+        for index in range(1, len(elements)):
+            elements[index - 1].link(elements[index])
+
         self.parse.set_property("num-channels", 1)
+        self.parse.set_property("sample-rate", SAMPLE_RATE)
+        self.caps_filter.set_property(
+            "caps",
+            Gst.caps_from_string(
+                f"audio/x-raw,format=S16LE,channels=1,rate={SAMPLE_RATE}"
+            ),
+        )
         self.pitch.set_property("pitch", 1)
 
         bus = self.pipeline.get_bus()
         bus.add_signal_watch()
         bus.connect("message::error", self.on_eos_or_end)
         bus.connect("message::eos", self.on_eos_or_end)
-        bus.connect("message::state-changed", self.on_state_changed)
-
-    def on_state_changed(self, bus, msg):
-        if msg.src != self.pipeline:
-            return
-
-        old, new, pending = msg.parse_state_changed()
-        if new == Gst.State.PLAYING:
-            self.emit("start")
 
     def on_eos_or_end(self, bus, msg):
         if msg.type == Gst.MessageType.ERROR:
@@ -73,40 +77,40 @@ class Mimic3SynthWorker(GObject.Object):
         self._reset_pipeline()
 
     def _reset_pipeline(self):
+        if not self.source:
+            return
         rr = os.fdopen(self.source.get_property("fd"), "rb")
         rr.flush()
         rr.close()
+        os.close(self.sink.get_property("fd"))
+
         self.source.set_state(Gst.State.NULL)
         self.source.unlink(self.parse)
         self.pipeline.remove(self.source)
         self.source = None
         self.pitch.set_property("pitch", 1)
         self.pipeline.set_state(Gst.State.READY)
-        self.emit("end")
+        self.emit("done")
 
     @GObject.Signal
-    def start(self):
+    def done(self):
         pass
 
-    @GObject.Signal
-    def end(self):
-        pass
-
-    def _synth(self, fd, text, voice_id, rate, volume):
+    def _synth(self, fd, text, voice_id, rate):
         ww = os.fdopen(fd, "wb", buffering=0)
-        with self._tts_lock:
+        with Mimic3SynthWorker._tts_lock:
             s = mimic3_tts.Mimic3Settings()
             if voice_id:
                 self.mimic3.voice = voice_id
             self.mimic3.rate = rate if rate is not None else s.rate
-            self.mimic3.volume = volume * 100 if volume is not None else s.volume
             self.mimic3.begin_utterance()
             self.mimic3.speak_text(text)
             for result in self.mimic3.end_utterance():
                 ww.write(result.audio_bytes)
         ww.close()
 
-    def speak(self, text, voice_id=None, pitch=None, rate=None, volume=None):
+    def synthesize(self, fd, text, voice_id=None, pitch=None, rate=None):
+        self.sink.set_property("fd", fd)
         self.source = Gst.ElementFactory.make("fdsrc", "source")
         self.pipeline.add(self.source)
         self.source.link(self.parse)
@@ -115,25 +119,15 @@ class Mimic3SynthWorker(GObject.Object):
         r, w = os.pipe()
         self.source.set_property("fd", r)
 
-        x = threading.Thread(target=self._synth, args=(w, text, voice_id, rate, volume))
+        x = threading.Thread(target=self._synth, args=(w, text, voice_id, rate))
         x.start()
 
         self.pipeline.set_state(Gst.State.PLAYING)
-
-    def pause(self):
-        self.pipeline.set_state(Gst.State.PAUSED)
-
-    def resume(self):
-        self.pipeline.set_state(Gst.State.PLAYING)
-
-    def stop(self):
-        self._reset_pipeline()
 
 
 class MimicProvider(dbus.service.Object):
     def __init__(self, loop, *args):
         self._last_speak_args = [0, "", "", 0, 0, 0]
-        self._tasks = {}
         self._loop = loop
         if not os.environ.get("KEEP_ALIVE"):
             GLib.timeout_add_seconds(AUTO_EXIT_SECONDS, self._timeout)
@@ -141,72 +135,34 @@ class MimicProvider(dbus.service.Object):
         self.mimic3 = mimic3_tts.Mimic3TextToSpeechSystem(s)
         self.mimic3.preload_voice(self.mimic3.voice)
         self._worker_pool = []
-        for i in range(1):
+        for i in range(5):
             worker = Mimic3SynthWorker(self.mimic3)
+            worker.connect("done", self._on_done)
             self._worker_pool.append(worker)
         dbus.service.Object.__init__(self, *args)
 
     def _timeout(self):
-        if self._tasks:
+        if len(self._worker_pool) < 5:
             return True
         self._loop.quit()
         return False
 
-    @dbus.service.method(
-        "org.freedesktop.Speech.Provider",
-        in_signature="tssddd",
-        out_signature="",
-    )
-    def Speak(self, task_id, utterance, voice_id, pitch, rate, volume):
-        worker = self._worker_pool.pop(0)
-        self._tasks[task_id] = worker
-        worker.connect("start", self.on_start, task_id)
-        worker.connect("end", self.on_end, task_id)
-        worker.speak(utterance, voice_id, pitch, rate, volume)
-
-    def on_end(self, worker, task_id):
-        worker.disconnect_by_func(self.on_end)
-        worker.disconnect_by_func(self.on_start)
+    def _on_done(self, worker):
         self._worker_pool.append(worker)
-        self.SpeechEnd(task_id)
-
-    def on_start(self, worker, task_id):
-        self.SpeechStart(task_id)
 
     @dbus.service.method(
         "org.freedesktop.Speech.Provider",
-        in_signature="t",
+        in_signature="hssdd",
         out_signature="",
     )
-    def Pause(self, task_id):
-        worker = self._tasks.get(task_id)
-        if worker:
-            worker.pause()
-
-    @dbus.service.method(
-        "org.freedesktop.Speech.Provider",
-        in_signature="t",
-        out_signature="",
-    )
-    def Resume(self, task_id):
-        worker = self._tasks.get(task_id)
-        if worker:
-            worker.resume()
-
-    @dbus.service.method(
-        "org.freedesktop.Speech.Provider",
-        in_signature="t",
-        out_signature="",
-    )
-    def Cancel(self, task_id):
-        worker = self._tasks.get(task_id)
-        if worker:
-            worker.stop()
+    def Synthesize(self, fd, utterance, voice_id, pitch, rate):
+        worker = self._worker_pool.pop(0)
+        worker.synthesize(fd.take(), utterance, voice_id, pitch, rate)
 
     @dbus.service.method(
         "org.freedesktop.Speech.Provider",
         in_signature="",
-        out_signature="a(ssas)",
+        out_signature="a(sssas)",
     )
     def GetVoices(self):
         voices = []
@@ -220,6 +176,7 @@ class MimicProvider(dbus.service.Object):
                         (
                             f"{v.name} - {lang_desc}",
                             v.key,
+                            f"audio/x-raw,format=S16LE,channels=1,rate={SAMPLE_RATE}",
                             [lang_tag],
                         )
                     )
@@ -229,22 +186,11 @@ class MimicProvider(dbus.service.Object):
                             (
                                 f"{v.name}/{speaker} - {lang_desc}",
                                 f"{v.key}#{speaker}",
+                                f"audio/x-raw,format=S16LE,channels=1,rate={SAMPLE_RATE}",
                                 [v.language],
                             )
                         )
         return voices
-
-    @dbus.service.signal("org.freedesktop.Speech.Provider", signature="t")
-    def SpeechStart(self, task_id):
-        pass
-
-    @dbus.service.signal("org.freedesktop.Speech.Provider", signature="ttt")
-    def SpeechRangeStart(self, task_id, start, end):
-        pass
-
-    @dbus.service.signal("org.freedesktop.Speech.Provider", signature="t")
-    def SpeechEnd(self, task_id):
-        pass
 
     @dbus.service.signal("org.freedesktop.Speech.Provider")
     def VoicesChanged(self):
